requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pyyaml>=6.0
